Refactoring to split the main script file into several classes
