+++ src/assets/style.css
@@ -12,17 +12,10 @@
     display: flex;
     align-items: center;
 }
-.spinner-border {
-    display: inline-block;
-    width: 30px;
-    height: 30px;
-    border: 4px solid transparent;
-    border-top-color: #007bff;
-    border-radius: 50%;
-    animation: spin 1s linear infinite;
-}
 #centralLoadingSpinner {
-    margin-right: 10px; /* Adjust as needed */
+    margin: 10px;
+    min-width: 16px!important;
+    min-height: 16px!important;
 }
 @keyframes spin {
     0% {
