+++ src/js/codelists-tab.js
@@ -1,0 +1,12 @@
+import { TabController } from "./tab-controller.js";
+
+export class CodelistsTab extends TabController {
+
+    constructor() {
+        super('codelists-tab');
+    }
+
+    async fetchAndRender() {
+        // TODO: implement
+    }
+}
