+++ src/js/config.js
@@ -1,0 +1,19 @@
+export const appConfig = {
+    tagsBaseUrl: 'https://api.github.com/repos/OP-TED/eForms-SDK',
+    rawBaseUrl: 'https://raw.githubusercontent.com/OP-TED/eForms-SDK',
+    noticeTypesFileUrl: 'https://api.github.com/repos/OP-TED/eForms-SDK/contents/notice-types'
+};
+
+export const domElements = {
+    loadingSpinner: $('#loadingSpinner'),
+    noticeTypesSpinner: $('#noticeTypesSpinner'),
+    apiStatus: $('#apiStatus'),
+    xmlStructureTree: $('#xmlStructureTree'),
+    tagsDropdown: $('#tagsDropdown'),
+    comparisonDropdown: $('#comparisonDropdown'),
+    noticeTypesDropdown: $('#noticeTypesDropdown'),
+    noticeTypesTree: $('#noticeTypesTree'),
+    fieldDetailsContent: $('#fieldDetailsContent'),
+    noticeTypesComparisonContent: $('#noticeTypesComparisonContent'),
+    noticeTypesDetails: $('#noticeTypesDetails')
+};
