+++ src/js/home-tab.js
@@ -1,0 +1,8 @@
+import { TabController } from "./tab-controller.js";
+
+export class HomeTab extends TabController {
+
+    constructor() {
+        super('home-tab');
+    }
+}
