+++ src/js/property-card.js
@@ -11,16 +11,14 @@ export class PropertyCard extends BootstrapWebComponent {
     }
 
     attributeChangedCallback(name, oldValue, newValue) {
-        if (oldValue === newValue) {
+        if (newValue === oldValue) {
             return;
         }
 
-        if (name === 'property-name') {
-            this.propertyName = newValue;
-        } else if (name === 'new-property-value') {
-            this.newPropertyValue = newValue;
-        } else if (name === 'old-property-value') {
-            this.oldPropertyValue = newValue;
+        switch (name) {
+            case 'property-name': this.propertyName = newValue; break;
+            case 'new-property-value': this.newPropertyValue = newValue; break;
+            case 'old-property-value': this.oldPropertyValue = newValue; break;
         }
 
         if (this.isConnected) {
@@ -48,6 +46,23 @@ export class PropertyCard extends BootstrapWebComponent {
         } else {
             this.shadowRoot.querySelector('#old-property-value').style.display = 'none';
         }    
+    }
+
+    static create(propertyName, currentValue, previousValue) {
+
+        const $component = $('<property-card/>');
+        $component.attr('property-name', propertyName + ': ');
+        $component.attr('new-property-value', PropertyCard.#formatPropertyValue(currentValue));
+        $component.attr('old-property-value', PropertyCard.#formatPropertyValue(previousValue));
+    
+        return $component;
+    }
+
+    static #formatPropertyValue(value) {
+        if (_.isObject(value) || Array.isArray(value)) {
+            return _.isObject(value) ? JSON.stringify(value, null, 2).replace(/\n/g, '<br>').replace(/ /g, '&nbsp;') : value;
+        }
+        return value;
     }
 }
