+++ src/js/schemas-tab.js
@@ -1,0 +1,12 @@
+import { TabController } from "./tab-controller.js";
+
+export class SchemasTab extends TabController {
+
+    constructor() {
+        super('schemas-tab');
+    }
+
+    async fetchAndRender() {
+        // TODO: implement
+    }
+}
