+++ src/js/state.js
@@ -1,0 +1,9 @@
+export const appState = {
+    sortedData: [],
+    versionData: [],
+    comparisonData: [],
+    sdkVersion: '',
+    baseVersion: '',
+    selectedNoticeTypeFile: 'notice-types.json',
+    activeTab: 'fields'
+};
