+++ src/js/translations-tab.js
@@ -1,0 +1,12 @@
+import { TabController } from "./tab-controller.js";
+
+export class TranslationsTab extends TabController {
+
+    constructor() {
+        super('translations-tab');
+    }
+
+    async fetchAndRender() {
+        // TODO: implement
+    }
+}
