+++ src/js/view-templates-tab.js
@@ -1,0 +1,12 @@
+import { TabController } from "./tab-controller.js";
+
+export class ViewTemplatesTab extends TabController {
+
+    constructor() {
+        super('view-templates-tab');
+    }
+
+    async fetchAndRender() {
+        // TODO: implement
+    }
+}
